Split toMatchSnapshot.test.js into 2 suits
