--- e2e/to-match-snapshot/package.json
@@ -1,5 +1,0 @@
-{
-  "jest": {
-    "testEnvironment": "node"
-  }
-}
